+++ kernels/volk/volk_32f_atan_32f.h
@@ -132,7 +132,7 @@ volk_32f_atan_32f_a_sse4_1(float* bVector, const float* aVector, unsigned int nu
 
   number = quarterPoints * 4;
   for(;number < num_points; number++){
-    *bPtr++ = atan(*aPtr++);
+    *bPtr++ = atanf(*aPtr++);
   }
 }
 
@@ -196,7 +196,7 @@ volk_32f_atan_32f_u_sse4_1(float* bVector, const float* aVector, unsigned int nu
 
   number = quarterPoints * 4;
   for(;number < num_points; number++){
-    *bPtr++ = atan(*aPtr++);
+    *bPtr++ = atanf(*aPtr++);
   }
 }
 
@@ -212,7 +212,7 @@ volk_32f_atan_32f_generic(float* bVector, const float* aVector, unsigned int num
   unsigned int number = 0;
 
   for(number = 0; number < num_points; number++){
-    *bPtr++ = atan(*aPtr++);
+    *bPtr++ = atanf(*aPtr++);
   }
 }
 #endif /* LV_HAVE_GENERIC */
